FEAT: user delete exception handling

+++ src/main/java/com/educandoweb/shopline/services/UserService.java
@@ -2,8 +2,11 @@ package com.educandoweb.shopline.services;
 
 import com.educandoweb.shopline.entities.User;
 import com.educandoweb.shopline.repositories.UserRepository;
+import com.educandoweb.shopline.services.exceptions.DatabaseException;
 import com.educandoweb.shopline.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,8 +31,16 @@ public class UserService {
         return this.userRepository.save(user);
     }
 
-    public void delete(Long id) {
-        this.userRepository.deleteById(id);
+    public void delete(Long id) throws ResourceNotFoundException {
+        if (!this.userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            this.userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id, User userData) {
